server/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.0
Flask-Migrate==4.0.1
SQLAlchemy==2.0.39
bcrypt==4.1.0
PyJWT==2.10.1
psycopg2-binary==2.9.10
python-dotenv==1.2.1
marshmallow==3.20.0
Flask-Cors==4.0.0
pydantic==2.9.0
gunicorn==21.2.0
apispec==6.8.4
email-validator==2.1.1
Faker==24.14.0
